% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{World Bank Global Population.}
\format{A data frame with 12035 rows and 8 variables:
\describe{
  \item{country_name}{Character, name of each country}
  \item{country_code}{Character, country identifier code}
  \item{indicator_name}{Character, WB indicator name}
  \item{indicator_code}{Character, WB indicator code}
  \item{year}{Numeric, year of observation}
  \item{population}{Numeric, population in millions}
  \item{region}{Character, World region}
  \item{incomegroup}{Character, WB income group}
  ...
}}
\source{
\url{http://https://data.worldbank.org/indicator/SP.POP.TOTL/}
}
\usage{
pop_data
}
\description{
A dataset containing the total population on country and regional level.
}
\keyword{datasets}
